requirements.txt
openai>=1.0.0
python-dotenv
requests
tqdm
numpy
scikit-learn
pandas
sentence-transformers
qdrant-client
